api/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
email-validator
python-multipart==0.0.6
pandas==1.5.3
numpy==1.24.3
groq
python-dotenv

# Google Cloud Storage
google-cloud-storage==2.10.0

# Optional: for production deployment
gunicorn==21.2.0
